Refactor feed store
