+++ src/client/bookmarks/Bookmarks.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
 import { connect } from 'react-redux';
 import { injectIntl, FormattedMessage } from 'react-intl';
-import { getFeed, getPendingBookmarks } from '../store/reducers';
+import { getPendingBookmarks } from '../store/reducers';
 import Feed from '../feed/Feed';
 import {
   getFeedFromState,
@@ -11,13 +11,14 @@ import {
   getFeedHasMoreFromState,
 } from '../helpers/stateHelpers';
 import { reload } from '../store/authStore/authActions';
-import { getBookmarks } from '../feed/feedActions';
+import { getBookmarks } from '../store/feedStore/feedActions';
 import { showPostModal } from '../store/appStore/appActions';
 import requiresLogin from '../auth/requiresLogin';
 import PostModal from '../post/PostModalContainer';
 import MobileNavigation from '../components/Navigation/MobileNavigation/MobileNavigation';
 import { getIsReloading } from '../store/authStore/authSelectors';
 import { getPosts } from '../store/postsStore/postsSelectors';
+import { getFeed } from '../store/feedStore/feedSelectors';
 
 @requiresLogin
 @injectIntl
