+++ src/client/object/ObjectFeed/ObjectFeed.js
@@ -5,17 +5,19 @@ import { FormattedMessage, injectIntl } from 'react-intl';
 import { isEmpty, uniq, get, filter } from 'lodash';
 import PropositionContainer from '../../rewards/Proposition/PropositionList/PropositionListContainer';
 import Feed from '../../feed/Feed';
-import { getFeed, getReadLanguages, getSuitableLanguage } from '../../store/reducers';
+import { getReadLanguages, getSuitableLanguage } from '../../store/reducers';
 import {
   getFeedLoadingFromState,
   getFeedHasMoreFromState,
   getFeedFromState,
 } from '../../helpers/stateHelpers';
-import { getObjectPosts, getMoreObjectPosts } from '../../feed/feedActions';
+import { getObjectPosts, getMoreObjectPosts } from '../../store/feedStore/feedActions';
 import { showPostModal } from '../../store/appStore/appActions';
 import PostModal from '../../post/PostModalContainer';
 import * as ApiClient from '../../../waivioApi/ApiClient';
 import Loading from '../../components/Icon/Loading';
+import { getFeed } from '../../store/feedStore/feedSelectors';
+
 import './ObjectFeed.less';
 
 @injectIntl
