+++ src/client/replies/Replies.js
@@ -3,14 +3,13 @@ import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
 import { injectIntl } from 'react-intl';
 import { connect } from 'react-redux';
-import { getFeed } from '../store/reducers';
 import {
   getFeedFromState,
   getFeedLoadingFromState,
   getFeedHasMoreFromState,
 } from '../helpers/stateHelpers';
 import { showPostModal } from '../store/appStore/appActions';
-import { getReplies, getMoreReplies } from '../feed/feedActions';
+import { getReplies, getMoreReplies } from '../store/feedStore/feedActions';
 import Feed from '../feed/Feed';
 import Loading from '../components/Icon/Loading';
 import Affix from '../components/Utils/Affix';
@@ -19,6 +18,7 @@ import RightSidebar from '../app/Sidebar/RightSidebar';
 import PostModal from '../post/PostModalContainer';
 import requiresLogin from '../auth/requiresLogin';
 import { getAuthenticatedUserName, getIsAuthenticated } from '../store/authStore/authSelectors';
+import { getFeed } from '../store/feedStore/feedSelectors';
 
 class Replies extends React.Component {
   static propTypes = {
