+++ src/client/store/feedStore/feedSelectors.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export const getFeed = state => state.feed;
