+++ src/client/user/UserProfile.js
@@ -4,7 +4,7 @@ import { isEmpty } from 'lodash';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 import Feed from '../feed/Feed';
-import { getFeed, getUsersAccountHistory, getUser } from '../store/reducers';
+import { getUsersAccountHistory, getUser } from '../store/reducers';
 import {
   getFeedLoadingFromState,
   getFeedFetchedFromState,
@@ -12,7 +12,7 @@ import {
   getFeedFromState,
 } from '../helpers/stateHelpers';
 import { getUserAccountHistory } from '../wallet/walletActions';
-import { getUserProfileBlogPosts } from '../feed/feedActions';
+import { getUserProfileBlogPosts } from '../store/feedStore/feedActions';
 import { showPostModal } from '../store/appStore/appActions';
 import EmptyUserProfile from '../statics/EmptyUserProfile';
 import EmptyUserOwnProfile from '../statics/EmptyUserOwnProfile';
@@ -23,6 +23,7 @@ import {
   getIsAuthenticated,
   isGuestUser,
 } from '../store/authStore/authSelectors';
+import { getFeed } from '../store/feedStore/feedSelectors';
 
 @withRouter
 @connect(
