+++ src/client/user/UserReblogs.js
@@ -3,19 +3,18 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import Feed from '../feed/Feed';
 
-import { getFeed } from '../store/reducers';
-
 import {
   getFeedLoadingFromState,
   getFeedHasMoreFromState,
   getFeedFromState,
 } from '../helpers/stateHelpers';
 import { showPostModal } from '../store/appStore/appActions';
-import { getFeedContent, getMoreFeedContent } from '../feed/feedActions';
+import { getFeedContent, getMoreFeedContent } from '../store/feedStore/feedActions';
 import PostModal from '../post/PostModalContainer';
 import EmptyUserProfile from '../statics/EmptyUserProfile';
 import EmptyUserOwnProfile from '../statics/EmptyUserOwnProfile';
 import { getAuthenticatedUser, getIsAuthenticated } from '../store/authStore/authSelectors';
+import { getFeed } from '../store/feedStore/feedSelectors';
 
 @connect(
   state => ({
